Move MotionPhoto logic in separate serivce
